few more changes: - change all instances of cmd to intptr_t (except for GAME_vmsyscall since that's all still 32-bit) - for 32-bit builds, replaced the qmm_import/qmm_export struct functions with naked functions that jmp to syscall/vmmain to be a bit more performant and smaller binary size (compiled only with QMM_JMP_STUBS macro)

+++ include/main.h
@@ -57,8 +57,8 @@ constexpr int QMM_FAIL_GAME_SHUTDOWN = 1;
 
 C_DLLEXPORT void dllEntry(eng_syscall_t syscall);	// initial entry point for non-GetGameAPI games
 C_DLLEXPORT void* GetGameAPI(void* import);			// initial entry point for GetGameAPI games
-C_DLLEXPORT intptr_t vmMain(int cmd, ...);
-intptr_t syscall(int cmd, ...);
+C_DLLEXPORT intptr_t vmMain(intptr_t cmd, ...);
+intptr_t syscall(intptr_t cmd, ...);
 
 constexpr bool QMM_DETECT_GETGAMEAPI = true;
 constexpr bool QMM_DETECT_DLLENTRY = false;
