+++ src/game_jamp.cpp
@@ -16,7 +16,7 @@ Created By:
 
 GEN_QMM_MSGS(JAMP);
 
-const char* JAMP_eng_msg_names(int cmd) {
+const char* JAMP_eng_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(G_PRINT);
 		GEN_CASE(G_ERROR);
@@ -353,7 +353,7 @@ const char* JAMP_eng_msg_names(int cmd) {
 	}
 }
 
-const char* JAMP_mod_msg_names(int cmd) {
+const char* JAMP_mod_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(GAME_INIT);
 		GEN_CASE(GAME_SHUTDOWN);
@@ -400,7 +400,7 @@ const char* JAMP_mod_msg_names(int cmd) {
 	}
 }
 
-bool JAMP_is_mod_trace_msg(int cmd) {
+bool JAMP_is_mod_trace_msg(intptr_t cmd) {
 	switch (cmd) {
 	case GAME_CLIENT_THINK:
 	case GAME_RUN_FRAME:
