+++ src/game_rtcwmp.cpp
@@ -16,7 +16,7 @@ Created By:
 
 GEN_QMM_MSGS(RTCWMP);
 
-const char* RTCWMP_eng_msg_names(int cmd) {
+const char* RTCWMP_eng_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(G_PRINT);
 		GEN_CASE(G_ERROR);
@@ -217,7 +217,7 @@ const char* RTCWMP_eng_msg_names(int cmd) {
 	}
 }
 
-const char* RTCWMP_mod_msg_names(int cmd) {
+const char* RTCWMP_mod_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(GAME_INIT);
 		GEN_CASE(GAME_SHUTDOWN);
@@ -238,7 +238,7 @@ const char* RTCWMP_mod_msg_names(int cmd) {
 	}
 }
 
-bool RTCWMP_is_mod_trace_msg(int cmd) {
+bool RTCWMP_is_mod_trace_msg(intptr_t cmd) {
 	switch (cmd) {
 	case GAME_CLIENT_THINK:
 	case GAME_RUN_FRAME:
