+++ src/game_rtcwsp.cpp
@@ -16,7 +16,7 @@ Created By:
 
 GEN_QMM_MSGS(RTCWSP);
 
-const char* RTCWSP_eng_msg_names(int cmd) {
+const char* RTCWSP_eng_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(G_PRINT);
 		GEN_CASE(G_ERROR);
@@ -220,7 +220,7 @@ const char* RTCWSP_eng_msg_names(int cmd) {
 	}
 }
 
-const char* RTCWSP_mod_msg_names(int cmd) {
+const char* RTCWSP_mod_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(GAME_INIT);
 		GEN_CASE(GAME_SHUTDOWN);
@@ -242,7 +242,7 @@ const char* RTCWSP_mod_msg_names(int cmd) {
 	}
 }
 
-bool RTCWSP_is_mod_trace_msg(int cmd) {
+bool RTCWSP_is_mod_trace_msg(intptr_t cmd) {
 	switch (cmd) {
 	case GAME_CLIENT_THINK:
 	case GAME_RUN_FRAME:
